【WA】[AtCoder Beginner Contest 128]D - equeue

+++ abc128/abc128_d.cpp
@@ -17,27 +17,27 @@ typedef tuple<string, int, int> T;
 int main() {
     int n, k;
     cin >> n >> k;
-    vector<int> dq(n);
+    vector<ll> dq(n);
     rep(i, n) {
         cin >> dq[i];
     }
-    int ans = 0;
-    for (int l = 0; l < k; l++) {
-        for (int r = 0; r < k - l; r++) {
+    ll ans = 0;
+    for (ll l = 0; l < k; l++) {
+        for (ll r = 0; r < k - l; r++) {
             if (l + r > n) break;
-            int d = k - l - r;
-            int now = 0;
-            vector<int> s;
-            for (int i = 0; i < l; i++) {
+            ll d = k - l - r;
+            ll now = 0;
+            vector<ll> s;
+            for (ll i = 0; i < l; i++) {
                 now += dq[i];
                 s.push_back(dq[i]);
             }
-            for (int i = n - r; i < n; i++) {
+            for (ll i = n - r; i < n; i++) {
                 now += dq[i];
                 s.push_back(dq[i]);
             }
             sort(s.begin(), s.end());
-            for (int i = 0; i < d; i++) {
+            for (ll i = 0; i < d; i++) {
                 if (i >= s.size()) break;
                 if (s[i] >= 0) break;
                 now -= s[i];
